Added saving of account locations on server shutdown. Changed to integer based health. Added an available bytes function to the packet input stream, to allow reverse compatibility. Added method to check if an entity type is the sub type of another type. Changed the read() methods of database IO so they read into preallocated objects, instead of allocating a new object. Fixed small regular expression error with a missing backslash in the server's entity definition loader.

+++ Common/src/com/pi/common/game/LivingEntity.java
@@ -10,14 +10,14 @@ public class LivingEntity extends Entity {
 	/**
 	 * Entity's current health.
 	 */
-	private float health = 1f;
+	private int health = 1;
 
 	/**
 	 * Sets an entity's current health.
 	 * 
 	 * @param sHealth the new health
 	 */
-	public final void setHealth(final float sHealth) {
+	public final void setHealth(final int sHealth) {
 		this.health = sHealth;
 	}
 
@@ -26,7 +26,7 @@ public class LivingEntity extends Entity {
 	 * 
 	 * @return the entity's health
 	 */
-	public final float getHealth() {
+	public final int getHealth() {
 		return health;
 	}
 }
